Simplify MSA package structure
